flaskblog/models.py
from datetime import datetime
from flask import current_app
from . import db, login_manager
from flask_login import UserMixin  # class that check an user status
# crete timed secret token
from itsdangerous import TimedJSONWebSignatureSerializer as Serializer

# help to load user from db
@login_manager.user_loader
def load_user(user_id):
    return User.query.get(int(user_id))


class User(db.Model, UserMixin):  # User table
    id = db.Column(db.Integer, primary_key=True)
    username = db.Column(db.String(20), unique=True, nullable=False)
    email = db.Column(db.String(120), unique=True, nullable=False)
    image_file = db.Column(db.String(60), nullable=False,
                           default='default.jpg')
    password = db.Column(db.String(60), nullable=False)  # will hash to 60 char
    posts = db.relationship('Post', backref='author',
                            lazy=True)  # relation to Post`

    def __repr__(self):
        return f"User('{self.username}', '{self.email}', '{self.image_file}')"

    def get_reset_token(self, expires_sec=1800):  # 30 min
        s = Serializer(current_app.config['SECRET_KEY'], expires_sec)
        return s.dumps({'user_id': self.id}).decode('utf-8')

    @staticmethod
    def verify_reset_token(token):
        s = Serializer(current_app.config['SECRET_KEY'])
        try:
            user_id = s.loads(token)['user_id']
        except:
            return None
        return User.query.get(user_id)


class Post(db.Model):  # Post table
    id = db.Column(db.Integer, primary_key=True)
    title = db.Column(db.String(100), unique=True, nullable=False)
    content = db.Column(db.Text, nullable=False)
    date_posted = db.Column(db.DateTime, nullable=False,
                            default=datetime.utcnow)
    # relation , one user -> many posts.
    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)

    def __repr__(self):
        return f"Post('{self.title}', '{self.date_posted}'"


# python shell :
# >>> from flaskblog import db
# >>> db.create_all()

# TODO
