flaskblog/requirements.txt
bcrypt==3.1.7
colorama==0.3.7
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
httplib2==0.9.2
Jinja2==2.11.2
Pillow==7.1.2
SQLAlchemy==1.3.16
Twisted==17.9.0
Werkzeug==1.0.1
WTForms==2.2.1

